+++ headers/track/stt/monocular_tool_tracker.hpp
@@ -7,7 +7,7 @@ namespace ttrk {
   class MonocularToolTracker : public SurgicalToolTracker {
 
   public:
-    MonocularToolTracker(const int radius, const int height, const std::string &calibration_filename);
+    MonocularToolTracker(const float  radius, const float  height, const std::string &calibration_filename);
     
     virtual ~MonocularToolTracker() {};
